requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
aiosqlite==0.19.0
gspread==5.12.4
google-auth==2.25.2